+++ src/types/apis/login.type.ts
@@ -1,0 +1,4 @@
+export interface LoginVo {
+  accessToken: string;
+  refreshToken: string;
+}
